+++ formats/src/main/scala/frawa/typesjson/formats/Formats.scala
@@ -14,11 +14,9 @@
  * limitations under the License.
  */
 
-package frawa.typedjson.eval
+package frawa.typedjson.formats
 
-import frawa.typedjson.parser.Value
 import frawa.typedjson.pointer.Pointer
-import frawa.typedjson.validation.{FormatMismatch, UnknownFormat}
 
 import java.net.URI
 import java.util.UUID
